This PR is in the same spirit of https://github.com/Ferrite-FEM/Ferrite.jl/pull/309 and not such a crowded catastrophe as https://github.com/Ferrite-FEM/Ferrite.jl/pull/307

I added `cellnodes!` and `cellcoords!` dispatches for `AbstractGrid` that only call the interface function of `<:AbstractGrid`. I noticed that this is needed when we did the first `<:AbstractGrid` prototype. I'm not sure if these definitions should better live in grid.jl. They assume a concrete homogeneous `cells` vector right now, since they are only called by `DofHandler`.

I'd like to do a follow up PR such that a `CellIterator` also can handle any <:AbstractGrid. Then, we would only need consistent usage of the interface functions and everything should work ootb for properly subtyped grids