@koehlerson 2022-01-07
Comparing heat equation on master vs this branch on my laptop (subject to a plethora of fluctuations) in order to see if there is a significant performance drawback. Note this is not intended as performance benchmark but rather as a rough estimate if something was utterly needed w.r.t. type instability

## PR
200 x 200 elements
julia> @time include("docs/src/examples/heat_equation.jl")
  1.148711 seconds (187.30 k allocations: 113.092 MiB, 2.27% gc time, 19.11% compilation time)
 
500 x 500 elements
julia> @time include("docs/src/examples/heat_equation.jl")
  6.478253 seconds (400.90 k allocations: 702.862 MiB, 4.93% gc time, 3.64% compilation time)
  
 ## Master
200 x 200 elements
julia> @time include("docs/src/examples/heat_equation.jl")
  1.152864 seconds (201.76 k allocations: 113.971 MiB, 11.94% gc time, 21.89% compilation time)
  
500 x 500 elements
julia> @time include("docs/src/examples/heat_equation.jl")
  6.244453 seconds (401.77 k allocations: 702.985 MiB, 3.77% gc time, 3.58% compilation time)